main functionality

+++ src/main/java/com/example/newsapi/service/NewsService.java
@@ -2,14 +2,12 @@ package com.example.newsapi.service;
 
 import com.example.newsapi.dto.NewsDTO;
 import com.example.newsapi.entity.News;
-import com.example.newsapi.exception.ResourceAlreadyExistsException;
 import com.example.newsapi.exception.ResourceNotFoundException;
 import com.example.newsapi.modelassembler.NewsModelAssembler;
 import com.example.newsapi.repository.NewsRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PagedResourcesAssembler;
 import org.springframework.hateoas.PagedModel;
@@ -36,9 +34,6 @@ public class NewsService {
     }
 
     public NewsDTO createNews(NewsDTO news){
-        if(newsRepository.existsById(news.getId()))
-            throw new ResourceAlreadyExistsException("News with id " + news.getId() + " already exists");
-
         return assembler.toModel(
                 newsRepository.save(assembler.toEntity(news))
         );
@@ -46,31 +41,25 @@ public class NewsService {
 
     public NewsDTO getNewsById(long id){
         return assembler.toModel(
-                newsRepository
-                        .findById(id)
+                newsRepository.findById(id)
                         .orElseThrow(()->new ResourceNotFoundException("Not found News with id " + id))
         );
     }
 
-    public NewsDTO updateNews(NewsDTO newNewsDTO, long id){
-        News newsToUpdate = newsRepository
-                .findById(id)
-                .orElseThrow(()->new ResourceNotFoundException("Not found News with id " + id));
-
-        News newNews = assembler.toEntity(newNewsDTO);
-        newsToUpdate.setTitle(newNews.getTitle());
-        newsToUpdate.setText(newNews.getText());
-        newsToUpdate.setDate(newNews.getDate());
-
+    public NewsDTO updateNews(NewsDTO requestedNewsDto, long id){
         return assembler.toModel(
-                newsRepository.save(newsToUpdate)
+                newsRepository.findById(id).map(news -> {
+                    news.setTitle(requestedNewsDto.getTitle());
+                    news.setText(requestedNewsDto.getText());
+                    news.setDate(requestedNewsDto.getDate());
+                    return newsRepository.save(news);
+                }).orElseThrow(()-> new ResourceNotFoundException("Not found News with id " + id))
         );
     }
 
-    public void deleteById(long id){
-        newsRepository
-                .findById(id)
-                .orElseThrow(()-> new ResourceNotFoundException("Not found News with id " + id));
+    public void deleteNewsById(long id){
+        if(!newsRepository.existsById(id))
+            throw new ResourceNotFoundException("Not found News with id " + id);
 
         newsRepository.deleteById(id);
     }
